Revert "merge genomeIntervalTree-test-case > master"
