+++ cerebra/check_coverage_loci.py
@@ -48,11 +48,9 @@ def vcf_to_dataframe(filename):
 def get_filenames():
 	""" get file names given path """
 	files = []
-
-	for file in os.listdir(cwd):
+	for file in os.listdir(cwd + "scVCF_filtered_all/"):
 		if file.endswith(".vcf"):
-			fullPath = cwd + file 
-
+			fullPath = cwd + 'scVCF_filtered_all/' + file 
 			files.append(fullPath)
     
 	return files
@@ -199,8 +197,7 @@ def are_hits_in_cosmic(queryList, SNP_bool):
 
 def build_genome_positions_dict(fileName):
 	""" creates dict with genome coords for cosmic filtered hits to specific GOI """
-
-	cell = fileName.replace(cwd, "")
+	cell = fileName.replace(cwd + "scVCF_filtered_all/", "")
 	cell = cell.replace(".vcf", "")	
 
 	df = vcf_to_dataframe(fileName)
@@ -305,8 +302,7 @@ def evaluate_coverage_driver(ROI_hits_dict, gene_, cd):
 		calls coverage_search_on_vcf() for each of those AA level hitsß """
 
 	for cell in ROI_hits_dict.keys():
-
-		vcf_path = cwd + cell + '.vcf'
+		vcf_path = cwd + 'scVCF_filtered_all/' + cell + '.vcf'
 		vcf = vcf_to_dataframe(vcf_path)
 
 		ROIs = ROI_hits_dict.get(cell)
@@ -360,37 +356,31 @@ def convert_to_df(cd):
 			if '?' not in mut:
 				df.loc[cell,mut] = cov
 
-	# alphabetize rows and cols
-	df = df.reindex(sorted(df.columns), axis=1)
-	df = df.reindex(sorted(df.index), axis=0)
-	df.index.name = 'sample'
-
 	return(df)
  
 
 
 """ get cmdline input """
 @click.command()
-@click.option('--genes_list', default = '/Users/lincoln.harris/code/cerebra/cerebra/wrkdir/genesList.csv', prompt='path to csv file with genes of interest to evaluate coverage for', required=True, type=str)
-@click.option('--nthread', default = 2, prompt='number of threads', required=True, type=int)
-@click.option('--outprefix', default = 'sampleOut.csv', prompt='prefix to use for outfile', required=True, type=str)
-@click.option('--vcf_dir', default = '/Users/lincoln.harris/code/cerebra/cerebra/wrkdir/vcf_test_set/', prompt='path to directory containing vcf files', required=True)
-@click.option('--cosmic_db', default = '/Users/lincoln.harris/code/cerebra/cerebra/wrkdir/CosmicGenomeScreensMutantExport.tsv', prompt='path to cosmic database', required=True)
+@click.option('--genes_list', default = 'genesList.csv', prompt='name of csv file with genes of interest to evaluate coverage for. should be in wrkdir', required=True, type=str)
+@click.option('--nthread', default = 16, prompt='number of threads', required=True, type=int)
+@click.option('--outprefix', default = 'sampleOut', prompt='prefix to use for outfile', required=True, type=str)
+@click.option('--wrkdir', default = '/Users/lincoln.harris/code/cerebra/cerebra/wrkdir/', prompt='s3 import directory', required=True)
  
 
 
-def check_coverage_loci(genes_list, nthread, outprefix, vcf_dir, cosmic_db):
+def check_coverage_loci(genes_list, nthread, outprefix, wrkdir):
 	""" evaluate coverage for each loci for which we find a variant, for 
 		a given set of genes  """
 	global genomePos_laud_db
 	global cosmic_genome_tree
 	global cwd
 
-	cwd = vcf_dir
+	cwd = wrkdir
 	fNames = get_filenames()
-	GOI_df = pd.read_csv(genes_list, header=None, names=['gene'])
+	GOI_df = pd.read_csv(cwd + genes_list, header=None, names=['gene'])
 	gene_names = list(GOI_df.gene)
-	database = pd.read_csv(cosmic_db, delimiter = '\t')
+	database = pd.read_csv(cwd + "CosmicGenomeScreensMutantExport.tsv", delimiter = '\t')
 	coverage_dict = {}
 
 	# driver loop 
@@ -400,9 +390,7 @@ def check_coverage_loci(genes_list, nthread, outprefix, vcf_dir, cosmic_db):
 		genomePos_laud_db = pd.Series(database_laud['Mutation genome position'])
 
 		# init interval tree
-		cosmic_genome_tree = utils.GenomeIntervalTree(
-            lambda row: utils.GenomePosition.from_str(str(row["Mutation genome position"])),
-            (record for idx, record in database_laud.iterrows()))
+		cosmic_genome_tree = utils.GenomeIntervalTree(lambda row: utils.GenomePosition.from_str(str(row["Mutation genome position"])), database_laud)
 
 		p = mp.Pool(processes=nthread)
 			
@@ -423,6 +411,7 @@ def check_coverage_loci(genes_list, nthread, outprefix, vcf_dir, cosmic_db):
 
 		coverage_dict = evaluate_coverage_driver(cells_dict_GOI_coords, gene, coverage_dict)
 	
+	print(coverage_dict)
 	coverage_df = convert_to_df(coverage_dict)
-	coverage_df.to_csv(outprefix)
-  
+	coverage_df.to_csv(cwd + 'coverage_df_test.csv')
+
