+++ cerebra/coverageBatch_v1.csv
@@ -1,0 +1,87 @@
+chrom,start_pos,end_pos,outfile
+17,7674241,7674241,TP53_S241F_coverageByCell.csv
+7,55161562,55161562,EGFR_R521K_coverageByCell.csv
+2,177234076,177234076,NFE2L2_G81S_coverageByCell.csv
+7,55174773,55174787,EGFR_E746_A750delELREA_coverageByCell.csv
+17,7674239,7674239,TP53_C242S_coverageByCell.csv
+2,177231362,177231362,NFE2L2_S414N_coverageByCell.csv
+13,32332592,32332592,BRCA2_N372H_coverageByCell.csv
+17,7676215,7676215,TP53_L52P_coverageByCell.csv
+17,7676154,7676154,TP53_P72R_coverageByCell.csv
+19,10492087,10492087,KEAP1_R272H_coverageByCell.csv
+2,177231085,177231085,NFE2L2_K506N_coverageByCell.csv
+12,25245351,25245351,KRAS_G12C_coverageByCell.csv
+7,55174771,55174785,EGFR_K745_A750>T_coverageByCell.csv
+17,7676040,7676040,TP53_Y110C_coverageByCell.csv
+19,4095413,4095413,MAP2K2_D341N_coverageByCell.csv
+19,10486751,10486751,KEAP1_S592R_coverageByCell.csv
+19,10994969,10994969,SMARCA4_R521W_coverageByCell.csv
+19,10994933,10994933,SMARCA4_A509S_coverageByCell.csv
+19,10499687,10499687,KEAP1_R116W_coverageByCell.csv
+7,55142322,55142322,EGFR_G42D_coverageByCell.csv
+19,10499609,10499609,KEAP1_T142M_coverageByCell.csv
+19,10491634,10491634,KEAP1_G423V_coverageByCell.csv
+19,4099224,4099224,MAP2K2_R299M_coverageByCell.csv
+7,55191822,55191822,EGFR_L858R_coverageByCell.csv
+19,10489732,10489732,KEAP1_R483H_coverageByCell.csv
+19,10492091,10492091,KEAP1_V271M_coverageByCell.csv
+17,7675088,7675088,TP53_V175L_coverageByCell.csv
+19,10989378,10989378,SMARCA4_G394W_coverageByCell.csv
+2,177231357,177231357,NFE2L2_H416Y_coverageByCell.csv
+19,10492052,10492052,KEAP1_Q284*_coverageByCell.csv
+19,10499811,10499811,KEAP1_Q75*_coverageByCell.csv
+17,7676576,7676576,TP53_A7V_coverageByCell.csv
+17,7675072,7675072,TP53_E180K_coverageByCell.csv
+19,4095421,4095421,MAP2K2_F338Y_coverageByCell.csv
+19,4123811,4123811,MAP2K2_P22S_coverageByCell.csv
+2,177231354,177231354,NFE2L2_V417L_coverageByCell.csv
+17,7676001,7676001,TP53_R123W_coverageByCell.csv
+17,7675175,7675175,TP53_G146W_coverageByCell.csv
+15,66435103,66435103,MAP2K1_F53L_coverageByCell.csv
+7,116699997,116699997,MET_K305*_coverageByCell.csv
+12,25245347,25245347,KRAS_G13D_coverageByCell.csv
+7,55152627,55152627,EGFR_A237V_coverageByCell.csv
+19,10491736,10491736,KEAP1_D389Y_coverageByCell.csv
+17,7674230,7674230,TP53_G245D_coverageByCell.csv
+2,177234232,177234232,NFE2L2_D29H_coverageByCell.csv
+19,10499685,10499685,KEAP1_E117K_coverageByCell.csv
+17,7673776,7673776,TP53_R282W_coverageByCell.csv
+19,11041479,11041479,SMARCA4_A1448V_coverageByCell.csv
+2,177234071,177234071,NFE2L2_E82D_coverageByCell.csv
+12,25225628,25225628,KRAS_A146P_coverageByCell.csv
+12,25227341,25227341,KRAS_Q61H_coverageByCell.csv
+13,32332343,32332343,BRCA2_N289H_coverageByCell.csv
+17,7676037,7676037,TP53_L111Q_coverageByCell.csv
+19,10499570,10499570,KEAP1_V155A_coverageByCell.csv
+19,10499463,10499463,KEAP1_A191T_coverageByCell.csv
+7,140753336,140753336,BRAF_V600E_coverageByCell.csv
+19,10499750,10499750,KEAP1_A95S_coverageByCell.csv
+17,7675148,7675148,TP53_F155V_coverageByCell.csv
+7,55205284,55205284,EGFR_R1100S_coverageByCell.csv
+1,162754686,162754686,DDR2_R83H_coverageByCell.csv
+4,55114137,55114137,KDR_P263T_coverageByCell.csv
+17,7675088,7675088,TP53_R175H_coverageByCell.csv
+19,10492201,10492201,KEAP1_R234P_coverageByCell.csv
+19,10489744,10489744,KEAP1_D479G_coverageByCell.csv
+19,4117495,4117495,MAP2K1_A76S_coverageByCell.csv
+19,10499451,10499451,KEAP1_G195V_coverageByCell.csv
+3,181713016,181713016,SOX2_G219V_coverageByCell.csv
+19,10499914,10499914,KEAP1_A40V_coverageByCell.csv
+19,10489270,10489270,KEAP1_W544C_coverageByCell.csv
+19,10499816,10499816,KEAP1_S73I_coverageByCell.csv
+14,61727632,61727632,HIF1a_M250I_coverageByCell.csv
+7,116700028,116700028,MET_N315S_coverageByCell.csv
+19,11030831,11030831,SMARCA4_G1162C_coverageByCell.csv
+19,10491812,10491812,KEAP1_G364D_coverageByCell.csv
+19,10492070,10492070,KEAP1_P278S_coverageByCell.csv
+3,41225412,41225412,CTNNB1_P192S_coverageByCell.csv
+19,10499535,10499535,KEAP1_V167F_coverageByCell.csv
+12,25227342,25227342,KRAS_Q61L_coverageByCell.csv
+8,116847688,116847688,RAD21_A570S_coverageByCell.csv
+19,10491647,10491647,KEAP1_G419W_coverageByCell.csv
+12,25225627,25225627,KRAS_A146V_coverageByCell.csv
+19,10499447,10499447,KEAP1_C196F_coverageByCell.csv
+3,41225206,41225206,CTNNB1_Q165H_coverageByCell.csv
+7,116699090,116699090,MET_K2N_coverageByCell.csv
+19,10492146,10492146,KEAP1_W252C_coverageByCell.csv
+19,10492231,10492231,KEAP1_S224Y_coverageByCell.csv
