+++ cerebra/coverageBatch_v3.csv
@@ -1,0 +1,88 @@
+chrom,start_pos,end_pos,outfile
+2,29209870,29209870,ALK_A1251D_coverageByCell.csv
+2,29532108,29532108,ALK_K321R_coverageByCell.csv
+8,116857381,116857381,RAD21_L192V_coverageByCell.csv
+8,116856698,116856698,RAD21_M254I_coverageByCell.csv
+8,116856253,116856253,RAD21_V284F_coverageByCell.csv
+8,116852642,116852642,RAD21_A410S_coverageByCell.csv
+7,55191819,55191819,EGFR_G857V_coverageByCell.csv
+7,55191776,55191776,EGFR_V843L_coverageByCell.csv
+7,55191741,55191741,EGFR_R831H_coverageByCell.csv
+7,116699377,116699377,MET_C98F_coverageByCell.csv
+5,31409105,31409105,DROSH_E1275*_coverageByCell.csv
+4,86750817,86750817,PTPN13_H1000L_coverageByCell.csv
+4,86689078,86689078,PTPN13_S145F_coverageByCell.csv
+4,55121164,55121164,KDR_P32T_coverageByCell.csv
+4,55104668,55104668,KDR_C654*_coverageByCell.csv
+4,55094856,55094856,KDR_A973S_coverageByCell.csv
+4,55089999,55089999,KDR_A1050V_coverageByCell.csv
+3,52407417,52407417,BAP1_A140V_coverageByCell.csv
+3,41239213,41239213,CTNNB1_M739I_coverageByCell.csv
+3,41225476,41225476,CTNNB1_C213F_coverageByCell.csv
+3,41224622,41224622,CTNNB1_S37F_coverageByCell.csv
+3,41224606,41224606,CTNNB1_D32N_coverageByCell.csv
+3,189864383,189864383,TP63_R244L_coverageByCell.csv
+2,177234237,177234237,NFE2L2_D27G_coverageByCell.csv
+2,177234217,177234217,NFE2L2_R34G_coverageByCell.csv
+2,177234216,177234216,NFE2L2_R34Q_coverageByCell.csv
+2,177234088,177234088,NFE2L2_D77H_coverageByCell.csv
+2,177234087,177234087,NFE2L2_D77G_coverageByCell.csv
+2,177234082,177234082,NFE2L2_E79Q_coverageByCell.csv
+2,177234081,177234081,NFE2L2_E79V_coverageByCell.csv
+2,177234078,177234078,NFE2L2_T80K_coverageByCell.csv
+2,177232579,177232579,NFE2L2_S136F_coverageByCell.csv
+2,177232558,177232558,NFE2L2_L143R_coverageByCell.csv
+2,177231174,177231174,NFE2L2_P477S_coverageByCell.csv
+19,1221968,1221968,STK11_G294C_coverageByCell.csv
+19,1221320,1221320,STK11_P281L_coverageByCell.csv
+19,1221230,1221230,STK11_G251V_coverageByCell.csv
+19,1220704,1220704,STK11_A241P_coverageByCell.csv
+19,1220429,1220429,STK11_G174R_coverageByCell.csv
+19,1207121,1207121,STK11_E70*_coverageByCell.csv
+19,1207085,1207085,STK11_G58R_coverageByCell.csv
+19,11058316,11058316,SMARCA4_E1496*_coverageByCell.csv
+19,11033438,11033438,SMARCA4_G1232V_coverageByCell.csv
+19,11033327,11033327,SMARCA4_Q1195H_coverageByCell.csv
+19,11030822,11030822,SMARCA4_L1161fs*3_coverageByCell.csv
+19,11030733,11030733,SMARCA4_T1129S_coverageByCell.csv
+19,11024424,11024424,SMARCA4_E1023Q_coverageByCell.csv
+19,10996250,10996250,SMARCA4_Q544P_coverageByCell.csv
+19,10994970,10994970,SMARCA4_R521P_coverageByCell.csv
+19,10987797,10987797,SMARCA4_Q331*_coverageByCell.csv
+19,10499990,10499990,KEAP1_R15L_coverageByCell.csv
+19,10499787,10499787,KEAP1_V83F_coverageByCell.csv
+19,10499735,10499735,KEAP1_L100P_coverageByCell.csv
+19,10499706,10499706,KEAP1_M110V_coverageByCell.csv
+19,10499571,10499571,KEAP1_V155F_coverageByCell.csv
+19,10499559,10499559,KEAP1_A159P_coverageByCell.csv
+19,10499551,10499551,KEAP1_M161I_coverageByCell.csv
+19,10499480,10499480,KEAP1_I185N_coverageByCell.csv
+19,10499455,10499455,KEAP1_Q193H_coverageByCell.csv
+19,10499423,10499423,KEAP1_R204P_coverageByCell.csv
+19,10499421,10499421,KEAP1_E205*_coverageByCell.csv
+19,10492211,10492211,KEAP1_L231V_coverageByCell.csv
+19,10492174,10492174,KEAP1_S243C_coverageByCell.csv
+19,10492164,10492164,KEAP1_F246L_coverageByCell.csv
+19,10491949,10491949,KEAP1_P318L_coverageByCell.csv
+19,10491944,10491944,KEAP1_R320W_coverageByCell.csv
+19,10491943,10491943,KEAP1_R320P_coverageByCell.csv
+19,10491908,10491908,KEAP1_G332C_coverageByCell.csv
+19,10491905,10491905,KEAP1_G333C_coverageByCell.csv
+19,10491659,10491659,KEAP1_R415C_coverageByCell.csv
+19,10489799,10489799,KEAP1_R460S_coverageByCell.csv
+19,10489770,10489770,KEAP1_R470H_coverageByCell.csv
+19,10489741,10489741,KEAP1_G480W_coverageByCell.csv
+19,10489732,10489732,KEAP1_R483C_coverageByCell.csv
+19,10489239,10489239,KEAP1_R554Q_coverageByCell.csv
+19,10486720,10486720,KEAP1_G603W_coverageByCell.csv
+19,10486694,10486694,KEAP1_E611D_coverageByCell.csv
+18,51065549,51065549,SMAD4_R361H_coverageByCell.csv
+17,7676570,7676570,TP53_C9*_coverageByCell.csv
+17,7670700,7670700,TP53_R337C_coverageByCell.csv
+14,61734187,61734187,HIF1a_K310N_coverageByCell.csv
+14,36519122,36519122,NKX2-1_S79L_coverageByCell.csv
+13,48362929,48362929,RB1_C278R_coverageByCell.csv
+13,48360098,48360098,RB1_M230I_coverageByCell.csv
+13,32398756,32398756,BRCA2_K3416fs*>3_coverageByCell.csv
+1,162772172,162772172,DDR2_K551N_coverageByCell.csv
+1,162770381,162770381,DDR2_R458H_coverageByCell.csv
