--- cerebra/lincoln_tests/test_cov_mod_out.csv
@@ -1,25 +1,0 @@
-sample,BRAF_Q257H,BRAF_V600E,BRAF_W450L,EGFR_L858R,KRAS_G12C,KRAS_G13C,TP53_P72R,TP53_S241F
-test0_kras_G13C,0:0,0:0,0:0,0:0,0:0,2:4,0:0,0:0
-test10_kras_G12C,0:0,0:0,0:0,0:0,10:10,0:0,0:0,0:0
-test11_kras_G12C,0:0,0:0,0:0,0:0,10:15,0:0,0:0,0:0
-test12_braf_W450L,0:0,0:0,2:4,0:0,0:0,0:0,0:0,0:0
-test13_braf_W450L,0:0,0:0,10:10,0:0,0:0,0:0,0:0,0:0
-test14_braf_W450L,0:0,0:0,10:15,0:0,0:0,0:0,0:0,0:0
-test15_braf_Q257H,2:4,0:0,0:0,0:0,0:0,0:0,0:0,0:0
-test16_braf_Q257H,10:10,0:0,0:0,0:0,0:0,0:0,0:0,0:0
-test17_braf_Q257H,10:15,0:0,0:0,0:0,0:0,0:0,0:0,0:0
-test18_tp53_S241F,0:0,0:0,0:0,0:0,0:0,0:0,0:0,2:4
-test19_tp53_S241F,0:0,0:0,0:0,0:0,0:0,0:0,0:0,10:10
-test1_kras_G13C,0:0,0:0,0:0,0:0,0:0,10:10,0:0,0:0
-test20_tp53_S241F,0:0,0:0,0:0,0:0,0:0,0:0,0:0,10:15
-test21_tp53_P72R,0:0,0:0,0:0,0:0,0:0,0:0,2:4,0:0
-test22_tp53_P72R,0:0,0:0,0:0,0:0,0:0,0:0,10:10,0:0
-test23_tp53_P72R,0:0,0:0,0:0,0:0,0:0,0:0,10:15,0:0
-test2_kras_G13C,0:0,0:0,0:0,0:0,0:0,10:15,0:0,0:0
-test3_egfr_L858R,0:0,0:0,0:0,2:4,0:0,0:0,0:0,0:0
-test4_egfr_L858R,0:0,0:0,0:0,10:10,0:0,0:0,0:0,0:0
-test5_egfr_L858R,0:0,0:0,0:0,10:15,0:0,0:0,0:0,0:0
-test6_braf_V600E,0:0,2:4,0:0,0:0,0:0,0:0,0:0,0:0
-test7_braf_V600E,0:0,10:10,0:0,0:0,0:0,0:0,0:0,0:0
-test8_braf_V600E,0:0,10:15,0:0,0:0,0:0,0:0,0:0,0:0
-test9_kras_G12C,0:0,0:0,0:0,0:0,2:4,0:0,0:0,0:0
